removing async_docker_swarm and refactoring old fixture
